**Order Summary** - product summary Cells  - product cell  - top cell  - user detail cell  - thank you cell - fixed price to be a Decimal datatype

+++ swift/ios-shoppe-demo/Model/Order.swift
@@ -16,8 +16,8 @@ class Order {
         items.filter { $0.title == productName }.first?.quantity += 1
     }
 
-    func cartOrderTotal() -> Double {
-        var cost = Double()
+    func cartOrderTotal() -> Decimal {
+        var cost = Decimal()
         items.forEach { (item) in
             cost += item.price.multiply(item.quantity)
         }
